Split the game module into multiple files
